\name{sparseMatrixColVars-package}

\alias{sparseMatrixColVars-package}
\alias{sparseMatrixColVars}
\docType{package}

\title{The sparseMatrixColVars package}

\description{
  The \pkg{sparseMatrixColVars} package provides an optimized \code{colVars()}
  for dgCMatrix objects implemented in C.

  See \code{?\link{dgCMatrix_colVars}}.
}

\keyword{package}
